# Faster AVX2 matrix multiplications for MoE models