--- a/llamafile/iqk_mul_mat_amd_avx2.cpp
+++ b/llamafile/iqk_mul_mat_amd_avx2.cpp
@@ -1,4 +1,3 @@
 #ifdef __x86_64__
-//#define iqk_mul_mat iqk_mul_mat_avx2
 #include "iqk_mul_mat.inc"
 #endif // __x86_64__