--- a/llamafile/sgemm.h
+++ b/llamafile/sgemm.h
@@ -14,6 +14,8 @@ bool iqk_mul_mat_moe(long, long, long, int, int, const void *, const void *,
         float *, long, long, const void *, int, int);
 bool iqk_mul_mat_moe_zen4(long, long, long, int, int, const void *, const void *,
         float *, long, long, const void *, int, int);
+bool iqk_mul_mat_moe_unsupported(long, long, long, int, int, const void *, const void *,
+        float *, long, long, const void *, int, int);
 
 bool llamafile_sgemm(long, long, long, const void *, long, const void *, long, void *, long, int,
                      int, int, int, int, int, int);