This PR is a follow up of PR's #394 and #405 and enables the faster matrix multiplications for legacy and k-quants introduced there also for MoE models.

The following table shows a prompt processing speed comparison between the main branch and this PR for Mixtral8x7B on a Ryzen-7950X CPU

| Quantization | PP-512 (main) | PP-512 (PR) | Speedup |
| ---:| ---:| ---:| ---:|
| Q4_0 |  59.2 | - | - |
| Q4_1 |  35.3 | 69.6 | 1.97 |
| Q5_0 |  30.6 | 65.4 | 2.14 |
| Q5_1 |  29.5 | 64.0 | 2.17 |
| Q2_K_S |  66.8 | 88.9 | 1.33 |
| Q3_K_S |  45.2 | 85.3 | 1.89 |
| Q4_K_S |  53.4 | 81.8 | 1.53 |
| Q5_K_S |  38.6 | 75.0 | 1.94 |
| Q6_K     |  41.8 | 85.6 | 2.05 |
|IQ4_XS   | 41.6 | 76.1 |  1.83 |