@jart 2024-05-21
On AMD Ryzen Threadripper PRO 7995WX with Mixtral 8x7b I'm seeing speedups for Q5\_K\_M as high as 2.6x. Some quick measurements on my end, for a context size of 20900 and a prompt size of 1611 tokens:

| quant    | tok/sec before | tok/sec after | speedup |
| ---      | --:            | --:           | --:     |
| Q2\_K    | 153.95         | 195.15        | 1.27x   |
| Q5\_K\_M | 121.16         | 314.70        | 2.60x   |

So once again, outstanding work!

P.S. I'm going to be looking into integrating the `llama-bench` command sometime soon.